+++ sqlalchemy_utils/expressions.py
@@ -1,5 +1,5 @@
 import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import JSON
+from sqlalchemy.dialects.postgresql import ARRAY, JSON
 from sqlalchemy.ext.compiler import compiles
 from sqlalchemy.sql.elements import ColumnClause
 from sqlalchemy.sql.expression import (
@@ -9,6 +9,7 @@ from sqlalchemy.sql.expression import (
     Executable,
     FunctionElement
 )
+from sqlalchemy.sql.functions import GenericFunction
 
 from sqlalchemy_utils.functions.orm import quote
 
@@ -92,13 +93,37 @@ def compile_array_get(element, compiler, **kw):
     )
 
 
-class row_to_json(FunctionElement):
+class row_to_json(GenericFunction):
     name = 'row_to_json'
     type = JSON
 
 
 @compiles(row_to_json, 'postgresql')
 def compile_row_to_json(element, compiler, **kw):
+    return "%s(%s)" % (element.name, compiler.process(element.clauses))
+
+
+class json_array_length(GenericFunction):
+    name = 'json_array_length'
+    type = sa.Integer
+
+
+@compiles(json_array_length, 'postgresql')
+def compile_json_array_length(element, compiler, **kw):
+    return "%s(%s)" % (element.name, compiler.process(element.clauses))
+
+
+class array_agg(GenericFunction):
+    name = 'array_agg'
+    type = ARRAY
+
+    def __init__(self, arg, **kw):
+        self.type = ARRAY(arg.type)
+        GenericFunction.__init__(self, arg, **kw)
+
+
+@compiles(array_agg, 'postgresql')
+def compile_json_array_length(element, compiler, **kw):
     return "%s(%s)" % (element.name, compiler.process(element.clauses))